+++ crates/ruma-client-api/CHANGELOG.md
@@ -11,8 +11,8 @@ Breaking changes:
   spec.
 - `BackupAlgorithm::MegolmBackupV1Curve25519AesSha2` is now a tuple variant
   containing a non-exhaustive struct.
-- The variants of `UserIdentifier` were changed to tuple variants containing a
-  non-exhaustive struct.
+- `UserIdentifier` can now represent an identifier with a custom type, and its
+  variants were changed to tuple variants containing a non-exhaustive struct.
   - `UserIdentifier::UserIdOrLocalpart` was renamed to `UserIdentifier::Matrix`.
 
 Improvements: